spike: support for compiling Typed Resources to `terraform-plugin-framework` in addition to `terraform-plugin-sdk@v2`
This proof-of-concept supports outputting Typed Data Sources and Resources for use with `terraform-plugin-framework` rather
than `terraform-plugin-sdk@v2` Data Sources and Resources.

At this point this is functional only for primitive fields (e.g. bool/float/int/string) and not the more complex objects
however this shows a means of implementing support for this.